% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{autocorr_dgp}
\alias{autocorr_dgp}
\title{Generates t*sum(ng) or t*G*ng random normal variables indexed by t and G}
\usage{
autocorr_dgp(t, G, ng, rho = 1, lambda = 1, gamma = 1, mu = 0,
  sigma = 1)
}
\arguments{
\item{t}{The number of time periods for all G}

\item{G}{The number of groups in each time period}

\item{ng}{Obs per group, either int or G-length vector}

\item{mu}{Mean for random normal distribution}

\item{sigma}{Standard deviation for ranfom normal distribution}
}
\value{
Matrix with random variable, t and G columns
}
\description{
Generates t*sum(ng) or t*G*ng random normal variables indexed by t and G
}

