% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{autocorr_group}
\alias{autocorr_group}
\title{Generate autocorrelated numbers by group}
\usage{
autocorr_group(t, G, rho = 1, mu = 0, sigma = 1)
}
\arguments{
\item{t}{The number of time periods}

\item{G}{The number of groups}

\item{rho}{Autocorrelation displacement parameter}

\item{mu}{Mean for random normal distribution}

\item{sigma}{Standard deviation for ranfom normal distribution}
}
\value{
Matrix with autocorrelated, t and G columns
}
\description{
Generate autocorrelated numbers by group
}

