% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{check_completed}
\alias{check_completed}
\title{Checks the number of rows already in a table with given id}
\usage{
check_completed(db, table, id, id_name = "id")
}
\arguments{
\item{db}{Database to check values in}

\item{table}{Table to check variable/value in}

\item{id}{ID value to look for}

\item{id_name}{Name of ID variable, default is "id"}
}
\value{
Number of rows in table with id
}
\description{
Checks the number of rows already in a table with given id
}

