% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_write_mc}
\alias{db_write_mc}
\title{Runs monte carlo and write results to database}
\usage{
db_write_mc(parameters, db, table, id_name = "id", reps, dgp, formula, bootby,
  clusterby, x_interest, H0 = 0, cores = 1)
}
\arguments{
\item{parameters}{List of DGP and bootstrap parameters}

\item{db}{Name of SQLite database}

\item{table}{Name of SQLite database}

\item{id_name}{Name of ID variable, default is "id"}

\item{reps}{Number of Monte Carlo replication}

\item{dgp}{DGP function}

\item{formula}{Formula for generating model from random data}

\item{bootby}{String or vector of strings indicating variables to bootstrap by}

\item{clusterby}{String or vector of strings indicating variables to cluster by}

\item{x_interest}{X variable of interest, defaults to 'X'}

\item{H0}{Number indicating the null hypothesis, default is 0}

\item{cores}{Number of cores for parallel bootstrap}

\item{id}{ID value to be added to each row}
}
\description{
Runs monte carlo and write results to database
}

