% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_write_param}
\alias{db_write_param}
\title{Writes parameter table to DB}
\usage{
db_write_param(db, table, parameters_list)
}
\arguments{
\item{db}{Name of SQLite database}

\item{table}{Name of SQLite database}

\item{parameters_list}{Named list of parameters}
}
\description{
Writes parameter table to DB
}

