% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_write_pvals}
\alias{db_write_pvals}
\title{Writes matrix of p-values to database}
\usage{
db_write_pvals(p_vals, db, table, id, id_name = "id")
}
\arguments{
\item{p_vals}{Named matrix or dataframe of p-values}

\item{db}{Name of SQLite database}

\item{table}{Name of SQLite database}

\item{id}{ID value to be added to each row}

\item{id_name}{Name of ID variable, default is "id"}
}
\description{
Writes matrix of p-values to database
}

