% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo_recur}
\alias{monte_carlo_recur}
\title{Repeat Monte Carlo repetitions}
\usage{
monte_carlo_recur(reps, stat_func, stat_args, rand_func, rand_args, progbar,
  lecuyer, cores, seed_ffwd)
}
\arguments{
\item{reps}{Integer indicating the total Monte Carlo repetitions}

\item{stat_func}{Function that takes a dataframe from rand_func and returns a vector of statistics}

\item{stat_args}{List of arguments to be passed to stat_func}

\item{rand_func}{Function that generates a dataframe of random data given rand_args}

\item{rand_args}{List of arguments to be passed to stat_func}

\item{progbar}{Progress bar object to be incremented}

\item{lecuyer}{Boolean indicating whether to use L'Ecuyer-style seed for multicore reproducibility}

\item{cores}{Integer indicating number of cores to use}

\item{seed_ffwd}{Integer indicating the number of L'Ecuyer streams to move ahead, for reproducibility purposes}
}
\value{
Vector or matrix of numerical statistics
}
\description{
Repeat Monte Carlo repetitions
}

