% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo_stat}
\alias{monte_carlo_stat}
\title{Run one Monte Carlo repetition}
\usage{
monte_carlo_stat(rep, stat_func, stat_args, rand_func, rand_args,
  progbar = NULL)
}
\arguments{
\item{rep}{Integer indicating the current Monte Carlo repetition}

\item{stat_func}{Function that takes a dataframe from rand_func and returns a vector of statistics}

\item{stat_args}{List of arguments to be passed to stat_func}

\item{rand_func}{Function that generates a dataframe of random data given rand_args}

\item{rand_args}{List of arguments to be passed to stat_func}

\item{progbar}{Progress bar object to be incremented}
}
\value{
Vector of numerical statistics
}
\description{
Run one Monte Carlo repetition
}

