% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{multicore_seeds_recur}
\alias{multicore_seeds_recur}
\title{Recursive function for creating list of needed L'Ecuyer seeds}
\usage{
multicore_seeds_recur(n, block, seed)
}
\arguments{
\item{n}{Integer indicating the current repetition}

\item{block}{Integer indicating the total number of reps}

\item{seed}{Initial seed either from current state or passed explicitely}
}
\value{
Seed for stream advanced block number times
}
\description{
Recursive function for creating list of needed L'Ecuyer seeds
}

