% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{multiway_DGP}
\alias{multiway_DGP}
\title{Generate multi-way grouped random data}
\usage{
multiway_DGP(num_dims, groups, rho = 0, theta = 1,
  dim_names = rev(LETTERS)[19:(18 + num_dims)])
}
\arguments{
\item{groups}{Vector of dims length with integers indicating total groups for each group dimension}

\item{rho}{Float indicating portion of first dimension groups for which cells wil have theta observations instead of 1}

\item{dims}{Integer indicating the number of group dimensions}
}
\value{
theta
}
\description{
Generate multi-way grouped random data
}

