% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{simple_DGP}
\alias{simple_DGP}
\title{Generates G*ng random normal variables indexed by G}
\usage{
simple_DGP(ng, G)
}
\arguments{
\item{ng}{Obs per group, either int or G-length vector}

\item{G}{The number of groups in each time period}
}
\value{
Dataframe with X, Y and clusterby random variables
}
\description{
Generates G*ng random normal variables indexed by G
}

