% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{wild_cluster_boot_bootby}
\alias{wild_cluster_boot_bootby}
\title{Run wild clustered bootstrap for one bootby variable}
\usage{
wild_cluster_boot_bootby(data, model, x_interest, clusterby, boot_dist,
  boot_reps, bootby = clusterby, H0 = 0, cores = 1)
}
\arguments{
\item{data}{Random data to perform wild clustered bootstrap on}

\item{model}{lm class model for bootstrapping on}

\item{x_interest}{X variable of interest, defaults to 'X'}

\item{clusterby}{String or vector of strings indicating variables to cluster by}

\item{boot_dist}{Vector of bootstrap distribution}

\item{boot_reps}{Integer for the number of bootstrap replications}

\item{bootby}{String or vector of strings indicating variables to bootstrap by}

\item{H0}{Number indicating the null hypothesis, default is 0}

\item{cores}{Number of cores for parallel bootstrap}
}
\value{
Vector with p-value for clusterby variable
}
\description{
Run wild clustered bootstrap for one bootby variable
}

