% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{wild_cluster_boot_mc}
\alias{wild_cluster_boot_mc}
\title{Run wild clustered bootstrap Monte Carlo and write to file}
\usage{
wild_cluster_boot_mc(dgp, dgp_args, formula, reps, x_interest, clusterby,
  boot_dist, boot_reps, bootby, H0 = 0, cores = 1, progbar = FALSE,
  save_type = "db", save_file = "wildclusterbootsim.db")
}
\arguments{
\item{dgp}{Function for generating random data}

\item{dgp_args}{Named list of arguments to be passed to dgp function}

\item{formula}{Formula for generating model from random data}

\item{x_interest}{X variable of interest, defaults to 'X'}

\item{clusterby}{String or vector of strings indicating variables to cluster by}

\item{boot_dist}{Vector of bootstrap distribution}

\item{boot_reps}{Integer for the number of bootstrap replications}

\item{bootby}{String or vector of strings indicating variables to bootstrap by}

\item{H0}{Number indicating the null hypothesis, default is 0}

\item{cores}{Number of cores for parallel bootstrap}

\item{progbar}{Boolean indicating whether to use progress bar}

\item{save_type}{Method for saving data, currently only supports db for SQLite}

\item{save_file}{String indicating file name to save to}

\item{Reps}{Number of total Monte Carlo repetitions}
}
\value{
Named matrix of p-values with row for eahch rep and columns for each combination of clusterby and bootby
}
\description{
Run wild clustered bootstrap Monte Carlo and write to file
}

