% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{wild_cluster_boot_rep}
\alias{wild_cluster_boot_rep}
\title{Run one replication of wild clustered bootstrap simulation}
\usage{
wild_cluster_boot_rep(dgp, dgp_args, formula, x_interest, clusterby, boot_dist,
  boot_reps, bootby = clusterby, H0 = 0, cores = 1, progbar = NULL, n)
}
\arguments{
\item{dgp}{Function for generating random data}

\item{dgp_args}{Named list of arguments to be passed to dgp function}

\item{formula}{Formula for generating model from random data}

\item{x_interest}{X variable of interest, defaults to 'X'}

\item{clusterby}{String or vector of strings indicating variables to cluster by}

\item{boot_dist}{Vector of bootstrap distribution}

\item{boot_reps}{Integer for the number of bootstrap replications}

\item{bootby}{String or vector of strings indicating variables to bootstrap by}

\item{H0}{Number indicating the null hypothesis, default is 0}

\item{cores}{Number of cores for parallel bootstrap}

\item{n}{Repition number}
}
\value{
Named vector of p-values, with names representing combination of clusterby and bootby
}
\description{
Run one replication of wild clustered bootstrap simulation
}

