% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{wildclustboot}
\alias{wildclustboot}
\title{Run wild clustered bootstrap for one bootby variable}
\usage{
wildclustboot(bootby, data, model, clusterby, boot_dist, boot_reps, x_interest,
  H0 = 0, cores = 1)
}
\arguments{
\item{bootby}{String or vector of strings indicating variables to bootstrap by}

\item{data}{Random data to perform wild clustered bootstrap on}

\item{model}{lm class model for bootstrapping on}

\item{clusterby}{String or vector of strings indicating variables to cluster by}

\item{boot_dist}{Vector of bootstrap distribution}

\item{boot_reps}{Integer for the number of bootstrap replications}

\item{x_interest}{X variable of interest, defaults to 'X'}

\item{H0}{Number indicating the null hypothesis, default is 0}
}
\value{
Vector with p-value for clusterby variable
}
\description{
Run wild clustered bootstrap for one bootby variable
}

