% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{wildclustboot_mc}
\alias{wildclustboot_mc}
\title{Run one replication of wild clustered bootstrap simulation}
\usage{
wildclustboot_mc(reps, dgp, dgp_args, formula, bootby, clusterby, boot_dist,
  boot_reps, x_interest, H0 = 0, cores = 1)
}
\arguments{
\item{dgp}{Function for generating random data}

\item{dgp_args}{Named list of arguments to be passed to dgp function}

\item{formula}{Formula for generating model from random data}

\item{bootby}{String or vector of strings indicating variables to bootstrap by}

\item{clusterby}{String or vector of strings indicating variables to cluster by}

\item{boot_dist}{Vector of bootstrap distribution}

\item{boot_reps}{Integer for the number of bootstrap replications}

\item{x_interest}{X variable of interest, defaults to 'X'}

\item{H0}{Number indicating the null hypothesis, default is 0}
}
\value{
Named matrix of p-values, with row for each rep and names representing combination of clusterby and bootby
}
\description{
Run one replication of wild clustered bootstrap simulation
}

