% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{dgpmc}
\alias{dgpmc}
\title{Run random data Monte Carlo}
\usage{
dgpmc(reps, stat_func, stat_args, rand_func, rand_args, names = NULL,
  progbar = FALSE)
}
\arguments{
\item{reps}{Integer indicating the total Monte Carlo repetitions}

\item{stat_func}{Function that takes a dataframe from rand_func and returns a vector of statistics}

\item{stat_args}{List of arguments to be passed to stat_func}

\item{rand_func}{Function that generates a dataframe of random data given rand_args}

\item{rand_args}{List of arguments to be passed to stat_func}

\item{progbar}{Progress bar object to be incremented}
}
\value{
Vector or matrix of numerical statistics
}
\description{
Run random data Monte Carlo
}

